requirements/nick.txt
# requirements.txt for nick's dev
django-debug-toolbar==1.2.1
django-sslserver